backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
groq==0.9.0
chromadb==0.5.0
httpcore==1.0.5
